# openapi/gofiftyville.yml
openapi: '3.0.3'
info:
  title: Gofiftyville API
  description: Fiftyville API written in Go.
  version: 1.0.0
  contact:
    url: 'https://gofityville.com/api/v1'
  x-logo:
    url: '../detective_golang.svg'
    altText: Gofiftyville Logo
  license:
    name: MIT
    url: 'https://opensource.org/license/mit/'
servers:
  - url: 'https://gofityville.com/api/v1'

paths:
  /crime-scene-reports:
    get:
      summary: Crime Scene Reports
      description: Get Crime Scene Reports
      operationId: getCrimeSceneReports
      parameters:
        - $ref: '#/components/parameters/Date'
        - $ref: '#/components/parameters/Street'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCrimeSceneReports'
              examples:
                default:
                  value:
                    - date: 'January 1, 2021'
                      street: 'Chamberlin Street'
                      description: 'Credit card fraud took place at 00:53. Two people witnessed the event.'
                    - date: 'January 6, 2021'
                      street: 'Aydede'
                      description: 'Reckless driving took place at 07:18. No known witnesses.'
                    - date: 'January 22, 2021'
                      street: 'Humphrey Lane'
                      description: 'Expired parking meter took place at 09:47. One person witnessed the incident.'

        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity

components:
  schemas:
    GetCrimeSceneReports:
      description: List of crime scene reports.
      type: array
      items:
        $ref: '#/components/schemas/CrimeSceneReport'
    CrimeSceneReport:
      description: Daily operating hours for the museum.
      type: object
      properties:
        date:
          type: string
          format: date
        street:
          type: string
        description:
          type: string
      required:
        - date
        - street
        - description

  parameters:
    Date:
      name: date
      in: query
      schema:
        type: string
        format: date
    Street:
      name: street
      in: query
      schema:
        type: string